+++ tests/test_hpc.py
@@ -1,4 +1,6 @@
 from ccbr_tools.pipeline.hpc import get_hpc
+from ccbr_tools.shell import shell_run
+import subprocess
 
 
 def test_hpc_biowulf():
@@ -27,3 +29,7 @@ def test_hpc_frce():
 def test_hpc_none():
     hpc = get_hpc(debug=" ")
     assert not any([hpc, hpc.name, *hpc.modules.values(), hpc.singularity_sif_dir])
+
+
+if __name__ == "__main__":
+    print(get_hpc(debug="biowulf"))